+++ src/pyaota/generator/document.py
@@ -70,11 +70,12 @@ class ExamDocument(Document):
             'QUESTIONSPACING': self.specs.get('questionspacing', '24pt'),
         }
         self.extra_preamble = self.specs.get('extra_preamble', '')
+        self.pagestyles = self.specs.get('pagestyles', DEFAULT_PAGESTYLES_TEMPLATE)
         logger.debug(f'Number of questions: {len(self.question_list)}')
         content = HEADMATTER
         if self.extra_preamble:
             content += '\n' + self.extra_preamble
-        content += '\n\n' + DEFAULT_PAGESTYLES_TEMPLATE + '\n\n' + BEGIN_DOCUMENT + '\n\n' + self.instructions + '\n\n'
+        content += '\n\n' + self.pagestyles + '\n\n' + BEGIN_DOCUMENT + '\n\n' + self.instructions + '\n\n'
         for q in self.question_list:
             logger.debug(f'Rendering question: {q["id"]} {q["type"]}')
             content += self.question_renderer(q).rstrip() + '\n\n'